requirements.txt
Flask
google-generativeai
redis
werkzeug
pyjwt